+++ enterprise/backend/src/baserow_enterprise/api/sso/saml/validators.py
@@ -2,7 +2,7 @@ import io
 
 from rest_framework import serializers
 from saml2.xml.schema import XMLSchemaError
-from saml2.xml.schema import validate as validate_saml
+from saml2.xml.schema import validate as validate_saml_metadata_schema
 
 from baserow_enterprise.sso.saml.exceptions import SamlProviderForDomainAlreadyExists
 from baserow_enterprise.sso.saml.models import SamlAuthProviderModel
@@ -24,7 +24,7 @@ def validate_unique_saml_domain(
 def validate_saml_metadata(value):
     metadata = io.StringIO(value)
     try:
-        validate_saml(metadata)
+        validate_saml_metadata_schema(metadata)
     except XMLSchemaError:
         raise serializers.ValidationError(
             "The metadata is not valid according to the XML schema."