+++ enterprise/backend/src/baserow_enterprise/sso/saml/handler.py
@@ -100,7 +100,7 @@ class SamlAuthProviderHandler:
     @classmethod
     def get_saml_auth_provider_from_saml_response(
         cls,
-        saml_raw_response: str,
+        raw_saml_response: str,
     ) -> SamlAuthProviderModel:
         """
         Parses the saml response and returns the authentication provider that needs to
@@ -116,13 +116,12 @@ class SamlAuthProviderHandler:
         """
 
         try:
-            decoded_saml_response = ElementTree.fromstring(
-                base64.b64decode(saml_raw_response).decode("utf-8")
-            )
-            issuer = decoded_saml_response.find(
+            saml_response = cls.decode_saml_response(raw_saml_response)
+            saml_response_xml_tree = ElementTree.fromstring(saml_response)
+            issuer = saml_response_xml_tree.find(
                 "{urn:oasis:names:tc:SAML:2.0:assertion}Issuer"
             ).text
-        except (binascii.Error, ElementTree.ParseError, AttributeError):
+        except (ElementTree.ParseError, AttributeError):
             raise InvalidSamlResponse("Impossible decode SAML response.")
 
         saml_auth_provider = SamlAuthProviderModel.objects.filter(
@@ -136,7 +135,8 @@ class SamlAuthProviderHandler:
     def get_user_identity_from_authn_response(
         cls,
         authn_response: AuthnResponse,
-    ) -> Dict[str, Any]:
+        saml_request_data: Dict[str, str],
+    ) -> UserInfo:
         """
         Extracts the user identity from the authn response and return a dict that
         can be sent to the UserHandler to create or update the user.
@@ -157,7 +157,7 @@ class SamlAuthProviderHandler:
             else:
                 last_name = ""
             name = f"{first_name} {last_name}".strip()
-        return UserInfo(email, name)
+        return UserInfo(email, name, **saml_request_data)
 
     @classmethod
     def get_saml_auth_provider_from_email(
@@ -192,13 +192,23 @@ class SamlAuthProviderHandler:
             raise InvalidSamlRequest("No valid SAML identity provider found.")
 
     @classmethod
-    def sign_in_user_from_saml_response(cls, saml_response: str) -> AbstractUser:
+    def decode_saml_response(self, raw_saml_response: str) -> str:
+        try:
+            return base64.b64decode(raw_saml_response).decode("utf-8")
+        except (binascii.Error, KeyError):
+            raise InvalidSamlResponse("Impossible decode SAML response.")
+
+    @classmethod
+    def sign_in_user_from_saml_response(
+        cls, saml_response: str, saml_request_data: Optional[Dict[str, str]] = None
+    ) -> AbstractUser:
         """
         Signs in the user using the SAML response received from the identity
         provider.
 
         :param saml_response: The encoded SAML response sent from the Identity
             Provider.
+        :param saml_request_data: The data that was sent in the SAML request.
         :raises InvalidSamlResponse: When the SAML response is not valid.
         :raises InvalidSamlConfiguration: When the SAML configuration is not
             valid.
@@ -221,7 +231,7 @@ class SamlAuthProviderHandler:
             )
             cls.check_authn_response_is_valid_or_raise(authn_response)
             idp_provided_user_info = cls.get_user_identity_from_authn_response(
-                authn_response
+                authn_response, saml_request_data or {}
             )
         except (InvalidSamlConfiguration, InvalidSamlResponse) as exc:
             raise exc
@@ -272,7 +282,7 @@ class SamlAuthProviderHandler:
             raise InvalidSamlConfiguration("No Location header found in SAML response.")
 
     @classmethod
-    def get_sign_in_url(cls, user_email: str, original_url: str = "") -> str:
+    def get_sign_in_url(cls, query_params: Dict[str, str]) -> str:
         """
         Returns the sign in url for the correct identity provider. This url is
         used to initiate the SAML authentication flow from the service provider.
@@ -285,7 +295,9 @@ class SamlAuthProviderHandler:
         :return: The redirect url to the identity provider.
         """
 
-        valid_relay_state_url = get_valid_frontend_url(original_url)
+        user_email = query_params.pop("email", None)
+        original_url = query_params.pop("original", "")
+        valid_relay_state_url = get_valid_frontend_url(original_url, query_params)
 
         try:
             saml_auth_provider = cls.get_saml_auth_provider_from_email(user_email)